% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAC.R
\name{AAC}
\alias{AAC}
\alias{AAC.sq_ami_bsc}
\title{Compute AAC matrix}
\usage{
AAC(x, ...)

\method{AAC}{sq_ami_bsc}(x, ..., NA_letter = getOption("tidysq_NA_letter"))
}
\arguments{
\item{x}{[\code{sq_ami_bsc}]\cr
An object this function is applied to.}

\item{...}{further arguments to be passed from or to other methods.}

\item{NA_letter}{[\code{character(1)}]\cr
A string that is used to interpret and display \code{NA} value in the
context of \code{\link[=sq-class]{sq class}}. Default value equals to
"\code{!}".}
}
\value{
A \code{\link[tibble]{tibble}} with number of rows equal to the
number of sequences and twenty columns. Each column has one-letter name equal
to one-letter code of respective amino acid and contains frequencies of this
amino acid in respective sequences from range [0, 1].
}
\description{
Computes AAC (Amino Acid Composition) matrix, which contains
frequencies of each of 20 basic amino acids for each sequence.
}
\details{
Frequency of an amino acid is understood as the number of times this amino
acid appears in the sequence divided by total number of elements of this
sequence.
}
\examples{
sq_ami <- tidysq::sq(c("PPAVMMFDILKKIQ", "PQEWYTWLPVMCTN"))
AAC(sq_ami)

}
