% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_weight.R
\name{molecular_weight}
\alias{molecular_weight}
\alias{molecular_weight.sq_ami_bsc}
\title{"expasy": uses "average" from https://web.expasy.org/findmod/findmod_masses.html#AA}
\usage{
molecular_weight(x, type = "expasy", ...)

\method{molecular_weight}{sq_ami_bsc}(
  x,
  type = "expasy",
  ...,
  NA_letter = getOption("tidysq_NA_letter")
)
}
\description{
"expasy": uses "average" from https://web.expasy.org/findmod/findmod_masses.html#AA
}
