% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseAAC.R
\name{PseAAC}
\alias{PseAAC}
\alias{PseAAC.sq_ami_bsc}
\title{Compute PseAAC matrix}
\usage{
PseAAC(x, feature_values, feature_name, ..., lambda = 50, weight_factor = 0.15)

\method{PseAAC}{sq_ami_bsc}(x, feature_values, feature_name, ..., lambda = 30, weight_factor = 0.15)
}
\arguments{
\item{x}{[\code{sq}]\cr
An object this function is applied to.}

\item{feature_values}{[\code{numeric}]\cr
A named vector, where names are letters and values are
a numerical representation of some amino acid property.}

\item{feature_name}{[\code{character(1)}]\cr
A name of used amino acid property.}

\item{lambda}{[\code{integer(1)}]\cr
Number of features to generate based on given property.}

\item{weight_factor}{[\code{numeric(1)}]\cr
Importance factor of single generated feature compared
to AAC features.}
}
\description{
Computes 20 + lambda features for given
sequences. Returns data in a \code{tibble} format. First
20 features are AAC-based amino acid frequencies, the
rest are based on correlation factors of specified
property.
}
\examples{
polarity_values <- c(
  A = -0.591, C = 1.343, D = 1.05, E = 1.357, F = -1.006,
  G = -0.384, H = 0.336, I = -1.239, K = 1.831, L = -1.019,
  M = -0.663, N = 0.945, P = 0.189, Q = 0.931, R = 1.538,
  S = -0.228, T = -0.032, V = -1.337, W = -0.595, Y = 0.26
)
sq_ami <- tidysq::sq(c("PPAVMMFDILKKIQ", "PQEWYTWLPVMCTN"))
PseAAC(sq_ami, polarity_values, "polarity",
       lambda = 10)

}
